+++ src/formatter.ts
@@ -25,25 +25,31 @@ export const runExternalFormatter = async (command: string, args: readonly strin
             if (!err) {
                 return reject(`Formatting of '${fileName}' aborted because file '${tempFilePath}' already exists.`);
             }
-            fs.writeFile(tempFilePath, fileContent, async (writeError) => {
+            fs.writeFile(tempFilePath, fileContent, (writeError) => {
                 if (writeError) {
                     return reject(`Formatting of '${fileName}' aborted because file '${tempFilePath}' could not be created: '${writeError.message}'.`);
                 }
-                child_process.execFile(command, args.concat(tempFilePath), (execError, execStdout, execStderr) => {
-                    fs.unlink(tempFilePath, (unlinkError) => {
-                        if (execError) {
-                            return reject(`Formatting of '${fileName}' failed (${execError.code}): '${execError.message}'.`);
-                        }
-                        // For now, we ignore this because the formatter may emit warnings which must not cause a failure.
-                        // It would probably be better to log it somehow.
-                        // if (execStderr) {
-                        //     return reject(`Formatting of '${fileName}' proceeded with an error: '${execStderr}'.`);
-                        // }
-                        if (unlinkError) {
-                            return reject(`Formatting of '${fileName}' ended with an error because file '${tempFilePath}' could not be deleted: '${unlinkError.message}'.`);
-                        }
-                        return resolve(execStdout);
+                child_process.execFile(command, ["-i", tempFilePath].concat(args), (execError, execStdout, execStderr) => {
+                    fs.readFile(tempFilePath, (readError, readData) => {
+                        fs.unlink(tempFilePath, (unlinkError) => {
+                            if (execError) {
+                                return reject(`Formatting of '${fileName}' failed (${execError.code}): '${execError.message}'.`);
+                            }
+                            if (readError) {
+                                return reject(`Formatting of '${fileName}' aborted because file '${tempFilePath}' could not be read: '${readError.message}'.`);
+                            }
+                            // For now, we ignore this because the formatter may emit warnings which must not cause a failure.
+                            // It would probably be better to log it somehow.
+                            // if (execStderr) {
+                            //     return reject(`Formatting of '${fileName}' proceeded with an error: '${execStderr}'.`);
+                            // }
+                            if (unlinkError) {
+                                return reject(`Formatting of '${fileName}' ended with an error because file '${tempFilePath}' could not be deleted: '${unlinkError.message}'.`);
+                            }
+                            return resolve(readData.toString('utf8'));
+                        });
                     });
+
                 });
             });
         });